pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "konvert"
version = "0.1.0"
description = "Universal data format converter CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "izag8216" }]
keywords = ["cli", "converter", "json", "yaml", "toml", "csv", "xml", "ini", "env", "developer-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "tomlkit>=0.12",
    "xmltodict>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
konvert = "konvert.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
